@@ -18,67 +18,66 @@
   "author": "Luis Rodrigues <luis.rodrigues@seedstarslabs.com>",
   "license": "MIT",
   "dependencies": {
-    "autoprefixer": "6.7.6",
-    "babel-core": "6.23.1",
-    "babel-loader": "6.3.2",
+    "autoprefixer": "7.1.2",
+    "babel-core": "6.26.0",
+    "babel-loader": "7.1.2",
     "babel-plugin-react-display-name": "2.0.0",
     "babel-plugin-react-transform": "2.0.2",
     "babel-plugin-transform-decorators-legacy": "1.3.4",
     "babel-plugin-undeclared-variables-check": "6.22.0",
-    "babel-polyfill": "6.23.0",
-    "babel-preset-es2015": "6.22.0",
+    "babel-polyfill": "6.26.0",
+    "babel-preset-es2015": "6.24.1",
     "babel-preset-es2015-loose": "8.0.0",
-    "babel-preset-react": "6.23.0",
-    "babel-preset-stage-1": "6.22.0",
-    "babel-runtime": "6.23.0",
-    "body-parser": "1.17.0",
-    "bootstrap-loader": "2.0.0-beta.22",
+    "babel-preset-react": "6.24.1",
+    "babel-preset-stage-1": "6.24.1",
+    "babel-runtime": "6.26.0",
+    "body-parser": "1.17.2",
+    "bootstrap-loader": "2.2.0",
     "bootstrap-sass": "3.3.7",
     "classnames": "2.2.5",
-    "css-loader": "0.26.2",
-    "dotenv": "2.0.0",
+    "clean-webpack-plugin": "0.1.16",
+    "css-loader": "0.28.5",
     "errorhandler": "1.5.0",
-    "es6-promise": "4.0.5",
+    "es6-promise": "4.1.1",
     "exports-loader": "0.6.4",
-    "extract-text-webpack-plugin": "2.0.0-beta.4",
-    "file-loader": "0.10.1",
+    "extract-text-webpack-plugin": "3.0.0",
+    "file-loader": "0.11.2",
     "font-awesome": "4.7.0",
     "font-awesome-webpack": "0.0.5-beta.2",
-    "history": "4.5.1",
-    "html-webpack-plugin": "2.28.0",
+    "history": "4.7.2",
+    "html-webpack-plugin": "2.30.1",
     "imports-loader": "0.7.1",
     "isomorphic-fetch": "2.2.1",
-    "jquery": "3.1.1",
+    "jquery": "3.2.1",
     "less": "2.7.2",
-    "less-loader": "2.2.3",
-    "method-override": "2.3.7",
-    "node-sass": "4.5.0",
-    "postcss-import": "9.1.0",
-    "postcss-loader": "1.3.3",
-    "react": "15.4.2",
-    "react-dom": "15.4.2",
-    "react-mixin": "3.0.5",
-    "react-redux": "4.4.6",
+    "less-loader": "4.0.5",
+    "node-sass": "4.5.3",
+    "postcss-import": "10.0.0",
+    "postcss-loader": "2.0.6",
+    "prop-types": "15.5.10",
+    "react": "15.6.1",
+    "react-dom": "15.6.1",
+    "react-mixin": "3.1.0",
+    "react-redux": "5.0.6",
     "react-router": "3.0.2",
     "react-router-redux": "4.0.8",
-    "redux": "3.6.0",
+    "redux": "3.7.2",
     "redux-thunk": "2.2.0",
-    "resolve-url-loader": "2.0.2",
-    "sass-loader": "4.1.0",
-    "style-loader": "0.13.2",
-    "tcomb-form": "0.9.13",
-    "url-loader": "0.5.8",
-    "webpack": "2.2.1",
-    "webpack-merge": "1.1.1",
+    "resolve-url-loader": "2.1.0",
+    "sass-loader": "6.0.6",
+    "style-loader": "0.18.2",
+    "tcomb-form": "0.9.17",
+    "url-loader": "0.5.9",
+    "webpack": "3.5.5",
+    "webpack-merge": "4.1.0",
     "whatwg-fetch": "2.0.3",
-    "yargs": "7.0.1"
+    "yargs": "8.0.2"
   },
   "devDependencies": {
     "babel-eslint": "7.2.3",
-    "chai": "3.5.0",
-    "chai-as-promised": "6.0.0",
-    "clean-webpack-plugin": "0.1.15",
-    "enzyme": "2.7.1",
+    "chai": "4.1.1",
+    "chai-as-promised": "7.1.1",
+    "enzyme": "2.9.1",
     "eslint": "4.5.0",
     "eslint-config-airbnb": "15.1.0",
     "eslint-plugin-import": "2.7.0",
