@@ -1,9 +1,7 @@
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const fontAwesomeConfig = require('./font-awesome.config');
+const buildExtractStylesLoader = require('./font-awesone-helper');
 
-fontAwesomeConfig.styleLoader = ExtractTextPlugin.extract({
-    fallbackLoader: 'style-loader',
-    loader: 'css-loader!less-loader'
-});
+fontAwesomeConfig.styleLoader = buildExtractStylesLoader(ExtractTextPlugin.extract(['css-loader', 'less-loader']));
 
 module.exports = fontAwesomeConfig;
