@@ -1,23 +1,30 @@
 /* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
-const jsdom = require('jsdom');
+const { JSDOM } = require('jsdom');
 const register = require('ignore-styles').default;
 
 register(['.scss', '.sass', '.css', '.png', '.jpeg']);
 
-const doc = jsdom.jsdom('<!doctype html><html><body></body></html>');
-const win = doc.defaultView;
+// jsdom config from: https://github.com/airbnb/enzyme/blob/master/docs/guides/jsdom.md
+const jsdom = new JSDOM('<!doctype html><html><body></body></html>');
+const { window } = jsdom;
 
-global.document = doc;
-global.window = win;
+function copyProps(src, target) {
+    const props = Object.getOwnPropertyNames(src)
+        .filter((prop) => {
+            return typeof target[prop] === 'undefined';
+        })
+        .map((prop) => {
+            return Object.getOwnPropertyDescriptor(src, prop);
+        });
+    Object.defineProperties(target, props);
+}
 
-// set width variable for tests or else App.spec.js will throw an NaN error
-global.document.body.clientWidth = 1024;
-
-Object.keys(window).forEach((key) => {
-    if (!(key in global)) {
-        global[key] = window[key];
-    }
-});
+global.window = window;
+global.document = window.document;
+global.navigator = {
+    userAgent: 'node.js',
+};
+copyProps(window, global);
 
 function mockStorage() {
     const storage = {};
@@ -43,3 +50,12 @@ function mockStorage() {
 
 global.localStorage = mockStorage();
 global.sessionStorage = mockStorage();
+
+// define btoa which is not part of jsdom
+// from: https://stackoverflow.com/questions/23097928/node-js-btoa-is-not-defined-error
+
+if (typeof btoa === 'undefined') {
+    global.btoa = (str) => {
+        return new Buffer(str).toString('base64');
+    };
+}
